ppc-symbolic: Factor register-related definitions into their own module
